feat: rename example
