+++ packages/IPadMultiScenesHeadlessExample/README.md
@@ -1,4 +1,4 @@
-# IPadNoMainSceneExample
+# IPadMultiScenesHeadlessExample
 
 ## Usage
