+++ packages/IPadMultiScenesHeadlessExample/app.json
@@ -1,0 +1,4 @@
+{
+  "name": "IPadMultiScenesHeadlessExample",
+  "displayName": "IPadMultiScenesHeadlessExample"
+}
