+++ packages/IPadMultiScenesHeadlessExample/babel.config.js
