+++ packages/IPadMultiScenesHeadlessExample/ios/IPadMultiScenesHeadlessExample.xcodeproj/xcshareddata/xcschemes/IPadNoMainSceneExample.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "IPadNoMainSceneExample.app"
-               BlueprintName = "IPadNoMainSceneExample"
-               ReferencedContainer = "container:IPadNoMainSceneExample.xcodeproj">
+               BuildableName = "IPadMultiScenesHeadlessExample.app"
+               BlueprintName = "IPadMultiScenesHeadlessExample"
+               ReferencedContainer = "container:IPadMultiScenesHeadlessExample.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "IPadNoMainSceneExampleTests.xctest"
-               BlueprintName = "IPadNoMainSceneExampleTests"
-               ReferencedContainer = "container:IPadNoMainSceneExample.xcodeproj">
+               BuildableName = "IPadMultiScenesHeadlessExampleTests.xctest"
+               BlueprintName = "IPadMultiScenesHeadlessExampleTests"
+               ReferencedContainer = "container:IPadMultiScenesHeadlessExample.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "IPadNoMainSceneExample.app"
-            BlueprintName = "IPadNoMainSceneExample"
-            ReferencedContainer = "container:IPadNoMainSceneExample.xcodeproj">
+            BuildableName = "IPadMultiScenesHeadlessExample.app"
+            BlueprintName = "IPadMultiScenesHeadlessExample"
+            ReferencedContainer = "container:IPadMultiScenesHeadlessExample.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "IPadNoMainSceneExample.app"
-            BlueprintName = "IPadNoMainSceneExample"
-            ReferencedContainer = "container:IPadNoMainSceneExample.xcodeproj">
+            BuildableName = "IPadMultiScenesHeadlessExample.app"
+            BlueprintName = "IPadMultiScenesHeadlessExample"
+            ReferencedContainer = "container:IPadMultiScenesHeadlessExample.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
