+++ packages/IPadMultiScenesHeadlessExample/ios/IPadMultiScenesHeadlessExample.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:IPadNoMainSceneExample.xcodeproj">
+      location = "group:IPadMultiScenesHeadlessExample.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
