+++ packages/IPadMultiScenesHeadlessExample/ios/IPadMultiScenesHeadlessExample/AppDelegate.mm
@@ -9,7 +9,7 @@
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
-  self.moduleName = @"IPadNoMainSceneExample";
+  self.moduleName = @"IPadMultiScenesHeadlessExample";
   // You can add your custom initial props in the dictionary below.
   // They will be passed down to the ViewController used by React Native.
   self.initialProps = @{};
