+++ packages/IPadMultiScenesHeadlessExample/ios/IPadMultiScenesHeadlessExample/Images.xcassets/AppIcon.appiconset/Contents.json
