+++ packages/IPadMultiScenesHeadlessExample/ios/IPadMultiScenesHeadlessExample/Images.xcassets/Contents.json
