+++ packages/IPadMultiScenesHeadlessExample/ios/IPadMultiScenesHeadlessExample/Info.plist
@@ -5,7 +5,7 @@
 	<key>CFBundleDevelopmentRegion</key>
 	<string>en</string>
 	<key>CFBundleDisplayName</key>
-	<string>IPadNoMainSceneExample</string>
+	<string>IPadMultiScenesHeadlessExample</string>
 	<key>CFBundleExecutable</key>
 	<string>$(EXECUTABLE_NAME)</string>
 	<key>CFBundleIdentifier</key>
