+++ packages/IPadMultiScenesHeadlessExample/ios/IPadMultiScenesHeadlessExampleTests/IPadMultiScenesHeadlessExampleTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface IPadNoMainSceneExampleTests : XCTestCase
+@interface IPadMultiScenesHeadlessExampleTests : XCTestCase
 
 @end
 
-@implementation IPadNoMainSceneExampleTests
+@implementation IPadMultiScenesHeadlessExampleTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
