+++ packages/IPadMultiScenesHeadlessExample/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "IPadNoMainSceneExample",
+  "name": "IPadMultiScenesHeadlessExample",
   "version": "0.6.0-rc.2",
   "private": true,
   "workspaces": {
