+++ packages/IPadMultiScenesHeadlessExample/patches/react-native-gesture-handler+2.12.0.patch
