--- packages/IPadNoMainSceneExample/app.json
@@ -1,4 +1,0 @@
-{
-  "name": "IPadNoMainSceneExample",
-  "displayName": "IPadNoMainSceneExample"
-}
